add auto approval
