+++ app/(api)/api/admin/approve-request/route.ts
@@ -1,7 +1,15 @@
 import { NextResponse } from "next/server";
 import { PrismaClient, SubmissionStatus } from "@prisma/client";
+import nodemailer from "nodemailer";
 
 const prisma = new PrismaClient();
+const transporter = nodemailer.createTransport({
+  service: "gmail",
+  auth: {
+    user: process.env.EMAIL_USER,
+    pass: process.env.EMAIL_PASS,
+  },
+});
 
 export async function POST(req: Request) {
   const { requestId } = await req.json();
@@ -32,28 +40,40 @@ export async function POST(req: Request) {
     );
   }
 
-  await prisma.candidate.update({
-    where: { id: candidate.id },
-    data: {
-      name: request.fullName,
-      website: request.website || null,
-      linkedin: request.linkedin || null,
-      city: request.city,
-      state: request.state,
-      position: request.position,
-      additionalNotes: request.additionalInfo || null,
-      email: request.email,
-      phone: request.phone,
-      status: SubmissionStatus.APPROVED,
-      verified: true,
-      clerkUserId: request.clerkUserId,
-    },
-  });
+  try {
+    await prisma.candidate.update({
+      where: { id: candidate.id },
+      data: {
+        name: request.fullName,
+        website: request.website || null,
+        linkedin: request.linkedin || null,
+        city: request.city,
+        state: request.state,
+        position: request.position,
+        additionalNotes: request.additionalInfo || null,
+        email: request.email,
+        phone: request.phone,
+        status: SubmissionStatus.APPROVED,
+        verified: true,
+        clerkUserId: request.clerkUserId,
+      },
+    });
 
-  await prisma.userValidationRequest.update({
-    where: { id: request.id },
-    data: { status: SubmissionStatus.APPROVED },
-  });
+    await prisma.userValidationRequest.update({
+      where: { id: request.id },
+      data: { status: SubmissionStatus.APPROVED },
+    });
+    // Notify admin of approval
+    await transporter.sendMail({
+      from: process.env.EMAIL_USER,
+      to: process.env.MY_EMAIL,
+      subject: `${request.fullName} Elevra profile is now verified`,
+      text: `${process.env.NEXT_PUBLIC_APP_URL}/candidate/${candidate.slug}`,
+    });
+  } catch (err) {
+    console.error(err);
+    return NextResponse.redirect(new URL("/candidate/verify/error", req.url));
+  }
 
   return NextResponse.redirect(new URL("/admin/user-validation", req.url));
 }
