+++ app/(api)/api/userValidationRequest/route.ts
@@ -3,85 +3,81 @@ import prisma from "@/prisma/prisma";
 import nodemailer from "nodemailer";
 
 export async function POST(request: Request) {
-  const body = await request.json();
-  const {
-    fullName,
-    email,
-    phone,
-    position,
-    website,
-    linkedin,
-    additionalInfo,
-    city,
-    state,
-    candidateId,
-    clerkUserId,
-    electionId,
-  } = body;
-
-  if (!candidateId || isNaN(Number(candidateId))) {
-    return NextResponse.json({ error: "Invalid candidateId" }, { status: 400 });
-  }
-  if (
-    !clerkUserId ||
-    typeof clerkUserId !== "string" ||
-    clerkUserId.trim() === ""
-  ) {
-    return NextResponse.json({ error: "Invalid clerkUserId" }, { status: 400 });
-  }
-  if (!electionId || isNaN(Number(electionId))) {
-    return NextResponse.json({ error: "Invalid electionId" }, { status: 400 });
-  }
-
-  const existingRequest = await prisma.userValidationRequest.findUnique({
-    where: { clerkUserId: clerkUserId || "" },
-  });
-
-  if (existingRequest) {
-    return NextResponse.json(
-      {
-        error: "A validation request for this user has already been submitted.",
-      },
-      { status: 400 }
-    );
-  }
-
-  const userValidationRequest = await prisma.userValidationRequest.create({
-    data: {
-      fullName,
-      email,
-      phone,
-      position,
-      website,
-      linkedin,
-      additionalInfo,
-      city,
-      state,
-      candidateId: Number(candidateId),
+  try {
+    const {
+      fullName = "",
+      email = "",
+      phone = "",
+      position = "",
+      website = "",
+      linkedin = "",
+      additionalInfo = "",
+      city = "",
+      state = "",
+      candidateId,
       clerkUserId,
-      electionId: Number(electionId),
-    },
-  });
+      electionId,
+    } = await request.json();
 
-  // Set up nodemailer transporter using your email service credentials
-  const transporter = nodemailer.createTransport({
-    service: "gmail", // or another service
-    auth: {
-      user: process.env.EMAIL_USER, // your email address
-      pass: process.env.EMAIL_PASS, // your email password or app password
-    },
-  });
+    if (!candidateId || isNaN(Number(candidateId))) {
+      return NextResponse.json(
+        { error: "Invalid candidateId" },
+        { status: 400 }
+      );
+    }
+    if (
+      !clerkUserId ||
+      typeof clerkUserId !== "string" ||
+      clerkUserId.trim() === ""
+    ) {
+      return NextResponse.json(
+        { error: "Invalid clerkUserId" },
+        { status: 400 }
+      );
+    }
+    if (!electionId || isNaN(Number(electionId))) {
+      return NextResponse.json(
+        { error: "Invalid electionId" },
+        { status: 400 }
+      );
+    }
 
-  // Define email options
-  const mailOptions = {
-    from: process.env.EMAIL_USER,
-    to: process.env.MY_EMAIL, // your email address to receive notifications
-    subject: `New Validation Request: ${userValidationRequest.fullName}`,
-    text: `Name: ${userValidationRequest.fullName}\nEmail: ${userValidationRequest.email}\nPhone: ${userValidationRequest.phone}}`,
-  };
+    const existingRequest = await prisma.userValidationRequest.findUnique({
+      where: { clerkUserId: clerkUserId },
+    });
 
-  // Send the email
-  await transporter.sendMail(mailOptions);
+    if (existingRequest) {
+      return NextResponse.json(
+        {
+          error:
+            "A validation request for this user has already been submitted.",
+        },
+        { status: 400 }
+      );
+    }
 
-  return NextResponse.json(userValidationRequest);
+    // Create a new validation request
+    const newRequest = await prisma.userValidationRequest.create({
+      data: {
+        fullName,
+        email,
+        phone,
+        position,
+        website,
+        linkedin,
+        additionalInfo,
+        city,
+        state,
+        candidateId: Number(candidateId),
+        clerkUserId,
+        electionId: Number(electionId),
+        status: "PENDING",
+      },
+    });
+
+    return NextResponse.json({ id: newRequest.id });
+  } catch (err) {
+    console.error("Error in POST /api/userValidationRequest:", err);
+    return NextResponse.json({ error: err }, { status: 500 });
+  }
 }
