+++ app/(voters)/candidate/verify/success/page.tsx
@@ -1,0 +1,5 @@
+import VerifySuccessClient from "./VerifySuccessClient";
+
+export default async function VerifySuccessPage() {
+  return <VerifySuccessClient />;
+}
