+++ package.json
@@ -51,6 +51,7 @@
     "zod": "^3.24.2"
   },
   "devDependencies": {
+    "@clerk/clerk-sdk-node": "^4.13.23",
     "@eslint/eslintrc": "^3",
     "@faker-js/faker": "^9.6.0",
     "@playwright/test": "^1.51.1",
